+++ NEWS.md
@@ -3,3 +3,10 @@
 NEW FEATURES
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# syntenet 0.99.5
+
+NEW FEATURES
+
+* Replaced distance measure for clustering of phylogenomic profiles: from
+vegan::vegdist to labdsv::dsvdis